fix geometry.
The previous approach does not work for double oblique or diagonal slices.
For diagonal slices, which direction is the main direction is not clear from the normal vector.
For (double) oblique slices, it is unclear what inplane rotation = 0 deg. exactly means.
It is definitely not the rotation around the vector orthogonal to main direction and slice normal. Probably it is results from a specific sequence of rotations, i.e. euler angles (S>C>T ...).
Unfortunately, which sequence that is is unclear, especially for multiple slice groups it is not obvious where to get the orientation string (SCT) for each slice group.

Thus, just parse the data first and use the quaternions specified in the MDBs.